# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Comment Sentiment Analyzer API
  description: |
    The Comment Sentiment Analyzer API provides comprehensive sentiment analysis for social media comments across multiple platforms. 
    This RESTful API enables content creators to connect their social media accounts, analyze comment sentiment, and gain actionable insights about their audience engagement.

    ## Features
    - Multi-platform social media integration (YouTube, Instagram, Twitter/X, TikTok)
    - AI-powered sentiment analysis with emotion detection
    - Theme clustering and keyword extraction
    - Real-time analysis progress tracking
    - Comprehensive export capabilities
    - Analysis comparison and trending

    ## Authentication
    The API uses JWT (JSON Web Token) based authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## Rate Limiting
    API requests are rate limited to ensure fair usage:
    - Authentication: 100 requests/hour
    - Platform Operations: 50 requests/hour  
    - Analysis Operations: 100 requests/hour
    - Export Operations: 20 requests/hour

  version: 1.0.0
  contact:
    name: Comment Sentiment Analyzer Support
    email: support@comment-analyzer.com
    url: https://docs.comment-analyzer.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.comment-analyzer.com/v1
    description: Production server
...
paths:
...
  /platforms/connect/{platform}:
    get:
      tags:
        - Platforms
      summary: Initiate OAuth connection for platform
      description: |
        Initiates OAuth flow for connecting a social media platform. 
        Returns authorization URL to redirect user to platform's OAuth consent screen.
      operationId: connectPlatform
      security:
        - BearerAuth: []
      parameters:
        - name: platform
          in: path
          required: true
          description: Social media platform to connect
          schema:
            $ref: "#/components/schemas/Platform"
          example: youtube
      responses:
        "200":
          description: OAuth URL generated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/OAuthInitiationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
...
  /platforms/posts:
    get:
      tags:
        - Posts
      summary: Fetch user posts from connected platforms
      description: |
        Retrieves posts from connected social media platforms with filtering and pagination support.
        Posts are fetched from platform APIs and cached for performance.
      operationId: getUserPosts
      security:
        - BearerAuth: []
      parameters:
        - name: platform
          in: query
          description: Filter by specific platform
          schema:
            $ref: "#/components/schemas/Platform"
        - name: limit
          in: query
          description: Number of posts to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: dateFrom
          in: query
          description: Start date filter (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          description: End date filter (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: search
          in: query
          description: Search query for post titles
          schema:
            type: string
            maxLength: 100
      responses:
        "200":
          description: Posts retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/PaginatedPostResponse"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "429":
          $ref: "#/components/responses/RateLimitError"
...
  /analysis/start:
    post:
      tags:
        - Analysis
      summary: Start sentiment analysis for a post
      description: |
        Initiates sentiment analysis for a specific post's comments. 
        Creates a background job and returns job ID for progress tracking.
        Analysis includes sentiment classification, emotion detection, theme extraction, and summary generation.
      operationId: startAnalysis
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartAnalysisRequest"
            examples:
              basic_analysis:
                summary: Basic analysis request
                value:
                  postId: "post_123"
              advanced_analysis:
                summary: Analysis with custom options
                value:
                  postId: "post_123"
                  options:
                    includeSpamDetection: true
                    includeToxicityFilter: true
                    maxComments: 1000
                    sentimentModel: "enhanced"
      responses:
        "201":
          description: Analysis started successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AnalysisJobResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "429":
          $ref: "#/components/responses/RateLimitError"
  /analysis/{jobId}/status:
    get:
      tags:
        - Analysis
      summary: Get analysis job status and progress
      description: |
        Returns current status and progress of an analysis job.
        Use this endpoint to poll for job completion. Job statuses include:
        pending, processing, completed, failed, cancelled.
      operationId: getAnalysisStatus
      security:
        - BearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          description: Analysis job identifier
          schema:
            type: string
      responses:
        "200":
          description: Job status retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AnalysisJobResponse"
        "401":
          $ref: "#/components/responses/AuthenticationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Platform:
      type: string
      enum: [YOUTUBE, INSTAGRAM, TWITTER, TIKTOK]
      description: Supported social media platforms
...
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Optional message providing additional context
        correlationId:
          type: string
          description: Request correlation ID for debugging
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
    ApiError:
      type: object
      required:
        - message
        - code
        - status
      properties:
        message:
          type: string
          description: Human-readable error message
        code:
          type: string
          description: Machine-readable error code
        status:
          type: integer
          description: HTTP status code
        details:
          type: object
          description: Additional error details
        correlationId:
          type: string
          description: Request correlation ID for debugging
        userMessage:
          type: string
          description: User-friendly error message
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Field-specific validation errors
    PaginationMeta:
      type: object
      required:
        - page
        - limit
        - total
        - totalPages
        - hasNext
        - hasPrev
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        totalPages:
          type: integer
          minimum: 0
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there are more items after current page
        hasPrev:
          type: boolean
          description: Whether there are items before current page
        nextCursor:
          type: string
          description: Cursor for next page (cursor-based pagination)
        prevCursor:
          type: string
          description: Cursor for previous page (cursor-based pagination)

    # Authentication Types
...
    OAuthInitiationResponse:
      type: object
      required:
        - authUrl
        - state
        - platform
      properties:
        authUrl:
          type: string
          format: uri
          description: OAuth authorization URL
          example: "https://accounts.google.com/oauth/authorize?client_id=..."
        state:
          type: string
          description: OAuth state parameter for CSRF protection
        platform:
          $ref: "#/components/schemas/Platform"
...
    PostResponse:
      type: object
      required:
        - id
        - platform
        - platformPostId
        - title
        - url
        - publishedAt
        - createdAt
        - hasAnalysis
      properties:
        id:
          type: string
          description: Unique post identifier
        platform:
          $ref: "#/components/schemas/Platform"
        platformPostId:
          type: string
          description: Platform-specific post ID
        title:
          type: string
          description: Post title or caption
        url:
          type: string
          format: uri
          description: Direct URL to the post
        publishedAt:
          type: string
          format: date-time
          description: Post publication timestamp
        createdAt:
          type: string
          format: date-time
          description: Post creation in our system timestamp
        thumbnailUrl:
          type: string
          format: uri
          description: Post thumbnail URL
        description:
          type: string
          description: Post description or excerpt
        viewCount:
          type: integer
          description: Number of views (if available)
        likeCount:
          type: integer
          description: Number of likes
        commentCount:
          type: integer
          description: Number of comments
        hasAnalysis:
          type: boolean
          description: Whether post has been analyzed
        latestAnalysisId:
          type: string
          description: Most recent analysis ID
    PaginatedPostResponse:
      type: object
      required:
        - items
        - pagination
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PostResponse"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"

    # Analysis Types
    StartAnalysisRequest:
      type: object
      required:
        - postId
      properties:
        postId:
          type: string
          description: Post ID to analyze
        options:
          $ref: "#/components/schemas/AnalysisOptions"
    AnalysisOptions:
      type: object
      properties:
        includeSpamDetection:
          type: boolean
          default: true
          description: Whether to include spam detection
        includeToxicityFilter:
          type: boolean
          default: true
          description: Whether to include toxicity filtering
        maxComments:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1000
          description: Maximum number of comments to analyze
        minCommentLength:
          type: integer
          minimum: 1
          default: 5
          description: Minimum comment length to include
        includeLanguages:
          type: array
          items:
            type: string
          description: Languages to include (ISO codes)
        includeReplies:
          type: boolean
          default: true
          description: Whether to analyze comment replies
        sentimentModel:
          type: string
          enum: [standard, enhanced, multilingual]
          default: standard
          description: Sentiment analysis model to use
    AnalysisJobResponse:
      type: object
      required:
        - id
        - postId
        - status
        - progress
        - currentStep
        - totalSteps
        - stepDescription
        - createdAt
      properties:
        id:
          type: string
          description: Unique job identifier
        postId:
          type: string
          description: Post being analyzed
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
          description: Job status
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Progress percentage
        currentStep:
          type: integer
          description: Current processing step
        totalSteps:
          type: integer
          description: Total number of steps
        stepDescription:
          type: string
          description: Description of current step
        createdAt:
          type: string
          format: date-time
          description: Job creation timestamp
        startedAt:
          type: string
          format: date-time
          description: Job start timestamp
        completedAt:
          type: string
          format: date-time
          description: Job completion timestamp
        error:
          type: string
          description: Error message (if failed)
        estimatedCompletionAt:
          type: string
          format: date-time
          description: Estimated completion time
        analysisResultId:
          type: string
          description: Analysis result ID (if completed)
...
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
          examples:
            validation_error:
              summary: Validation error example
              value:
                success: false
                error:
                  message: "Validation failed"
                  code: "VALIDATION_ERROR"
                  status: 400
                  fieldErrors:
                    email: ["Email is required", "Email format is invalid"]
                    password: ["Password must be at least 8 characters"]
    AuthenticationError:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
          examples:
            auth_error:
              summary: Authentication error
              value:
                success: false
                error:
                  message: "Authentication required"
                  code: "AUTHENTICATION_ERROR"
                  status: 401
                  userMessage: "Please log in to continue"
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ApiResponse"
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    $ref: "#/components/schemas/ApiError"
      headers:
        X-RateLimit-Limit:
          description: Request limit per time window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Time when rate limit resets (Unix timestamp)
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
...